explore-assistant-cloud-run/requirements.txt
fastapi
uvicorn[standard]
requests
google-auth
httpx